Feature - Divided Calendar Code into many Components
